+++ src/Hook/EntityHooks.php
@@ -13,6 +13,19 @@ use Drupal\farm_grazing_plan\Bundle\GrazingEvent;
 class EntityHooks {
 
   /**
+   * Implements hook_entity_type_alter().
+   */
+  #[Hook('entity_type_alter')]
+  public function entityTypeAlter(array &$entity_types) {
+
+    // Add a constraint to log entities to restrict editing fields that are
+    // managed by a grazing plan.
+    if (isset($entity_types['log'])) {
+      $entity_types['log']->addConstraint('GrazingEventLog');
+    }
+  }
+
+  /**
    * Implements hook_entity_bundle_info_alter().
    */
   #[Hook('entity_bundle_info_alter')]